Generate handle type for structs
