+++ cs-bindgen-shared/src/bindgen_struct.rs
@@ -1,5 +1,6 @@
+use proc_macro2::Span;
 use serde::*;
-use syn::ItemStruct;
+use syn::{spanned::Spanned, Error, Ident, ItemStruct};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct BindgenStruct {
@@ -7,11 +8,25 @@ pub struct BindgenStruct {
 }
 
 impl BindgenStruct {
-    pub fn from_item(_item: ItemStruct) -> syn::Result<Self> {
-        todo!()
+    pub fn from_item(item: ItemStruct) -> syn::Result<Self> {
+        // Return an error for generic structs.
+        if !item.generics.params.is_empty() {
+            return Err(Error::new(
+                item.generics.span(),
+                "Generic structs are not not supported with `#[cs_bindgen]`",
+            ));
+        }
+
+        Ok(Self {
+            ty_ident: item.ident.to_string(),
+        })
     }
 
     pub fn raw_ident(&self) -> &str {
-        todo!()
+        &self.ty_ident
+    }
+
+    pub fn ident(&self) -> Ident {
+        Ident::new(&self.ty_ident, Span::call_site())
     }
 }
